Refactor search hook in referral page
