+++ client/hooks/api/referral/search-referral.tsx
@@ -1,8 +1,16 @@
+import { ChangeEvent, useState } from "react"
 import apiService from "@/utils/common/api"
+import { postSortingOptions } from "@/utils/common/sorting/post"
+import { referralSortingOptions } from "@/utils/common/sorting/referer"
 import { useInfiniteQuery } from "@tanstack/react-query"
 
 import { QueryKeyString } from "@/types/common/query-key-string"
 import { ReferralType } from "@/types/common/referral-type"
+import useGetIndustryList from "@/hooks/api/industry/get-Industry-list"
+import useGetCityList from "@/hooks/api/location/get-city-list"
+import useGetCountryList from "@/hooks/api/location/get-country-list"
+import useGetProvinceList from "@/hooks/api/location/get-province-list"
+import useDebounce from "@/hooks/common/debounce"
 
 interface ISearchRefererFilterMeta {
   companyName: string
@@ -15,16 +23,87 @@ interface ISearchRefererFilterMeta {
   yoeMin?: string
 }
 
-const useSearchReferral = (
-  sorting: string,
-  filterMeta: ISearchRefererFilterMeta,
-  type: ReferralType
-) => {
+const useSearchReferral = (type: ReferralType) => {
   const keyString =
     type === ReferralType.REFEREE
       ? QueryKeyString.SEARCH_REFEREE
       : QueryKeyString.SEARCH_REFERRER
-  return useInfiniteQuery({
+
+  const [companyName, setCompanyName] = useState("")
+  const [provinceUuid, setProvinceUuid] = useState<undefined | string>()
+  const [countryUuid, setCountryUuid] = useState<undefined | string>()
+  const [cityUuid, setCityUuid] = useState<undefined | string>()
+  const [industryUuid, setIndustryUuid] = useState<undefined | string>()
+  const [yoeMin, setYoeMin] = useState<undefined | string>("0")
+  const [yoeMax, setYoeMax] = useState<undefined | string>("100")
+  const [sorting, setSorting] = useState(referralSortingOptions[0].value)
+  const debouncedCompanyName = useDebounce(companyName, 800)
+
+  const handleCompanyChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setCompanyName(e.target.value)
+  }
+
+  const handleCountryChange = (value: string) => {
+    setCountryUuid(value)
+  }
+  const handleProvinceChange = (value: string) => {
+    setProvinceUuid(value)
+  }
+  const handleCityChange = (value: string) => {
+    setCityUuid(value)
+  }
+
+  const handleIndustryChange = (value: string) => {
+    setIndustryUuid(value)
+  }
+
+  const handleSortingChange = (value: string) => {
+    setSorting(value)
+  }
+
+  const handleYeoMinChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const rawValue = e.target.value
+
+    // Parse the input value to an integer
+    const integerValue = parseInt(rawValue)
+
+    // Check if the parsed value is a valid integer
+    if (!isNaN(integerValue) && integerValue >= 0) {
+      // If it's a non-negative integer, set the value as is
+      setYoeMin(integerValue.toString())
+    } else {
+      // If it's negative or not a valid integer, set it to '0'
+      setYoeMin("0")
+    }
+  }
+
+  const handleYeoMaxChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const rawValue = e.target.value
+
+    // Parse the input value to an integer
+    const integerValue = parseInt(rawValue)
+
+    // Check if the parsed value is a valid integer
+    if (!isNaN(integerValue) && integerValue >= 0) {
+      // If it's a non-negative integer, set the value as is
+      setYoeMax(integerValue.toString())
+    } else {
+      // If it's negative or not a valid integer, set it to '0'
+      setYoeMax("0")
+    }
+  }
+
+  const filterMeta = {
+    companyName: debouncedCompanyName,
+    cityUuid,
+    countryUuid,
+    industryUuid,
+    provinceUuid,
+    sorting,
+    yoeMin,
+    yoeMax,
+  }
+  const result = useInfiniteQuery({
     queryKey: [keyString, { sorting, filterMeta, type }],
     queryFn: apiService.searchReferral,
     refetchOnMount: false,
@@ -37,6 +116,25 @@ const useSearchReferral = (
       }
     },
   })
+
+  return {
+    result,
+    handleCompanyChange,
+    handleCountryChange,
+    handleProvinceChange,
+    handleCityChange,
+    handleSortingChange,
+    handleIndustryChange,
+    handleYeoMinChange,
+    handleYeoMaxChange,
+    provinceUuid,
+    cityUuid,
+    countryUuid,
+    industryUuid,
+    yoeMax,
+    yoeMin,
+    sorting,
+  }
 }
 
 export default useSearchReferral
